+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 [#326](https://github.com/vmware/go-vcloud-director/pull/326)
 * Update VDC dynamic func to handle API version 35.0 [#327](https://github.com/vmware/go-vcloud-director/pull/327)
 * Added methods `vm.UpdateVmCpuAndMemoryHotAdd` and `vm.UpdateVmCpuAndMemoryHotAddAsyc` [#324](https://github.com/vmware/go-vcloud-director/pull/324)
+* Introduce low level OpenAPI client functions `OpenApiGetAllItems`,`OpenApiPostItemSync`,`OpenApiPostItemAsync`,
+`OpenApiPostItem`, `OpenApiGetItem`, `OpenApiPutItem`, `OpenApiPutItemSync`, `OpenApiPutItemAsync`,
+`OpenApiDeleteItem`, `OpenApiIsSupported`, `OpenApiBuildEndpoints`
+[#325](https://github.com/vmware/go-vcloud-director/pull/325)
 
 ## 2.8.0 (June 30, 2020)